[TASK] Extract AbstractQueryBuilder from QueryBuilder
This pr:

* Extracts all non TYPO3 related logic from the QueryBuilder into the AbstractQueryBuilder

Fixes: #2080
