+++ Classes/Domain/Search/Query/ParameterBuilder/FieldCollapsing.php
@@ -24,10 +24,8 @@ namespace ApacheSolrForTypo3\Solr\Domain\Search\Query\ParameterBuilder;
  *  This copyright notice MUST APPEAR in all copies of the script!
  ***************************************************************/
 
-use ApacheSolrForTypo3\Solr\Domain\Search\Query\ParameterBuilder\ParameterBuilder;
-use ApacheSolrForTypo3\Solr\Domain\Search\Query\QueryBuilder;
+use ApacheSolrForTypo3\Solr\Domain\Search\Query\AbstractQueryBuilder;
 use ApacheSolrForTypo3\Solr\System\Configuration\TypoScriptConfiguration;
-use Solarium\QueryType\Select\Query\Query;
 
 /**
  * The FieldCollapsing ParameterProvider is responsible to build the solr query parameters
@@ -142,10 +140,10 @@ class FieldCollapsing extends AbstractDeactivatable implements ParameterBuilder
     }
 
     /**
-     * @param QueryBuilder $parentBuilder
-     * @return QueryBuilder
+     * @param AbstractQueryBuilder $parentBuilder
+     * @return AbstractQueryBuilder
      */
-    public function build(QueryBuilder $parentBuilder): QueryBuilder
+    public function build(AbstractQueryBuilder $parentBuilder): AbstractQueryBuilder
     {
         $query = $parentBuilder->getQuery();
         if(!$this->getIsEnabled()) {
