+++ Classes/Domain/Search/Query/ParameterBuilder/Highlighting.php
@@ -24,6 +24,7 @@ namespace ApacheSolrForTypo3\Solr\Domain\Search\Query\ParameterBuilder;
  *  This copyright notice MUST APPEAR in all copies of the script!
  ***************************************************************/
 
+use ApacheSolrForTypo3\Solr\Domain\Search\Query\AbstractQueryBuilder;
 use ApacheSolrForTypo3\Solr\Domain\Search\Query\QueryBuilder;
 use ApacheSolrForTypo3\Solr\System\Configuration\TypoScriptConfiguration;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
@@ -178,10 +179,10 @@ class Highlighting extends AbstractDeactivatable implements ParameterBuilder
 
 
     /**
-     * @param QueryBuilder $parentBuilder
-     * @return QueryBuilder
+     * @param AbstractQueryBuilder $parentBuilder
+     * @return AbstractQueryBuilder
      */
-    public function build(QueryBuilder $parentBuilder): QueryBuilder
+    public function build(AbstractQueryBuilder $parentBuilder): AbstractQueryBuilder
     {
         $query = $parentBuilder->getQuery();
         if(!$this->getIsEnabled()) {
@@ -208,6 +209,5 @@ class Highlighting extends AbstractDeactivatable implements ParameterBuilder
         }
 
         return $parentBuilder;
-
     }
 }
