+++ Classes/Domain/Search/Query/ParameterBuilder/ParameterBuilder.php
@@ -25,8 +25,7 @@ namespace ApacheSolrForTypo3\Solr\Domain\Search\Query\ParameterBuilder;
  *  This copyright notice MUST APPEAR in all copies of the script!
  ***************************************************************/
 
-use ApacheSolrForTypo3\Solr\Domain\Search\Query\QueryBuilder;
-
+use ApacheSolrForTypo3\Solr\Domain\Search\Query\AbstractQueryBuilder;
 
 /**
  * The implementation of ParameterBuilder is responsible to build an array with
@@ -39,8 +38,8 @@ interface ParameterBuilder {
 
 
     /**
-     * @param QueryBuilder $parentBuilder
-     * @return QueryBuilder
+     * @param AbstractQueryBuilder $parentBuilder
+     * @return AbstractQueryBuilder
      */
-    public function build(QueryBuilder $parentBuilder): QueryBuilder;
+    public function build(AbstractQueryBuilder $parentBuilder): AbstractQueryBuilder;
 }
