+++ Classes/Domain/Search/Query/ParameterBuilder/Sortings.php
@@ -23,8 +23,8 @@ namespace ApacheSolrForTypo3\Solr\Domain\Search\Query\ParameterBuilder;
  *
  *  This copyright notice MUST APPEAR in all copies of the script!
  ***************************************************************/
-use TYPO3\CMS\Core\Utility\GeneralUtility;
 
+use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
  * The Sorting ParameterProvider is responsible to build the solr query parameters
